Style: norm 놈 맞춤
