+++ expansion/print_env.c
@@ -6,103 +6,13 @@
 /*   By: jungslee <jungslee@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/02 10:07:27 by jungslee          #+#    #+#             */
-/*   Updated: 2024/06/02 10:09:38 by jungslee         ###   ########.fr       */
+/*   Updated: 2024/06/02 10:34:16 by jungslee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "expansion.h"
 
-char	*env_strcpy(int start, int end, char *str)
-{
-	char	*ret;
-	int		i;
-
-	i = 0;
-	ret = (char *)malloc(sizeof(char) * (end - start + 2));
-	if (ret == NULL)
-		handle_error("exit : malloc error", 1, 0);
-	while (start + i <= end)
-	{
-		ret[i] = str[start + i];
-		i++;
-	}
-	ret[i] = '\0';
-	return (ret);
-}
-
-void	env_cpy(t_env **env, char **envp)
-{
-	char	*name;
-	char	*content;
-	int		i;
-	int		j;
-	int		equal_idx;
-
-	i = 0;
-	equal_idx = 0;
-	while (envp[i] != NULL)
-	{
-		j = 0;
-		while (envp[i][j] != '=')
-			j++;
-		equal_idx = j;
-		name = env_strcpy(0, j - 1, envp[i]);
-		content = ft_strdup(envp[i] + j + 1);
-		if (name == NULL || content == NULL)
-			handle_error("exit : malloc error", 1, 0);
-		env_add_back(env, env_new(name, content));
-		i++;
-	}
-}
-
-#include <stdio.h>
-
-void	print_all_env(t_env *head)//TODO ㅈㅣ우ㅓ
-{
-	t_env	*to_print;
-	int		i;
-
-	to_print = head;
-	i = 1;
-	while (to_print != NULL)
-		to_print = to_print->next;
-}
-
-int	env_strncmp(char *s1, char *name, int n)
-{
-	while (n-- != 0 || *name != '\0')
-	{
-		if (*s1 != *name)
-			return (1);
-		s1++;
-		name++;
-	}
-	if (*name != '\0')
-		return (1);
-	return (0);
-}
-
-int	is_alpha_num(char *var)
-{
-	int	i;
-	int	quote_flag;
-
-	i = 0;
-	quote_flag = 0;
-	while (var[i] != '\0')
-	{
-		if (!(ft_isalnum(var[i]) || var[i] == '_' || var[i] == '\"'))
-			return (0);
-		if (var[i] == '\"')
-			quote_flag = 1;
-		i++;
-	}
-	if (quote_flag == 1)
-		return (2);
-	return (1);
-}
-
-char	*replace_env(char *str, int start, int end, char *content)//TODO
+char	*replace_env(char *str, int start, int end, char *content)
 {
 	char	*ret;
 	char	*tmp1;
@@ -150,7 +60,6 @@ char	*is_valid_env(char *ori, int idx, t_env *env)
 	return (ret);
 }
 
- 
 void	is_var(t_node *node, t_env *env)
 {
 	int		i;
@@ -166,8 +75,9 @@ void	is_var(t_node *node, t_env *env)
 			if (node->cmd[i][j] == '$')
 			{
 				if (node->cmd[i][j + 1] == '$')
-					handle_error("해당 쉘의 pid 출력해야하는데 방법이 없음.", 1, 0);//TODO 고쳐
-				else if (!(node->cmd[i][j + 1] == '\"' || node->cmd[i][j + 1] == '\0'))
+					handle_error("해당 쉘의 pid 출력해야하는데 방법이 없음.", 1, 0);//TODO 의논해
+				else if (!(node->cmd[i][j + 1] == '\"' || \
+						node->cmd[i][j + 1] == '\0'))
 					node->cmd[i] = is_valid_env(node->cmd[i], j, env);
 			}
 			j++;
@@ -176,7 +86,7 @@ void	is_var(t_node *node, t_env *env)
 	}
 }
 
-void	check_env(t_node *ast, t_env *env)//TODO exit status 구현하기
+void	check_cmd_node(t_node *ast, t_env *env)//TODO exit status 구현하기
 {
 	t_node	*node;
 	char	*var;
@@ -185,8 +95,9 @@ void	check_env(t_node *ast, t_env *env)//TODO exit status 구현하기
 	if (node->type == N_CMD)
 	{
 		is_var(node, env);
+		is_wild_card(node);
 		return ;
 	}
-	check_env(node->right, env);
-	check_env(node->left, env);
+	check_cmd_node(node->right, env);
+	check_cmd_node(node->left, env);
 }
