+++ inc/expansion.h
@@ -6,7 +6,7 @@
 /*   By: jungslee <jungslee@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/30 17:15:23 by jungslee          #+#    #+#             */
-/*   Updated: 2024/05/31 21:58:23 by jungslee         ###   ########.fr       */
+/*   Updated: 2024/06/02 10:33:33 by jungslee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,12 @@ t_env	*env_new(char *name, char *content);
 t_env	*env_last(t_env *head);
 void	env_add_back(t_env **head, t_env *new);
 int		env_free_all(t_env **head);
-void	check_env(t_node *ast, t_env *env);
+void	check_cmd_node(t_node *ast, t_env *env);
 void	env_cpy(t_env **env, char **envp);
+char	*env_strcpy(int start, int end, char *str);
+int		env_strncmp(char *s1, char *name, int n);
+int		is_alpha_num(char *var);
+
 
 void	print_all_env(t_env *head); //TODO ㅈㅣ우ㅓㅜ
 #endif
